Added module with behavioural experiment
Added a module that looks at a second study we plan to make, comparing the behaviour between people who have and have not taken a particular teaching module. The module is far from finished, but I have gotten it to the stage where it does not promptly crash, and the numbers that come out look vaguely plausible.