Started moving methods from study to learning_module
I have started moving the manipulations from the study class to the learning_module class. The code is currently in a stage where it runs and passes my testing, but be aware that it is currently quite messy, as I have simply cut&pasted a large amount of code from one module to another and made some minor changes to it. I plan to refactor things a lot, since there are a bunch of workarounds that are no longer working around anything, since they were originally due to me splitting up tasks between the learning_module (then participants) class and the study class in a silly way.