Started changing file format to json
I initially used an ad hoc file format of my own for saving and loading data, since I was annoyed with how json cannot handle numpy arrays. That was a bad idea, given that my file format did not do any better, and in any case turning ndarrays into lists and back is not that hard. It started to be a problem once it became clear that I would likely need to do a lot of reading data, since we want to see the results for each skill and session per participant. So far, I have implemented json within the learning_module class.