Moved simulation method into simulated_learning_module
Since I am in the process of moving the handling of manipulations from the study class into the learning_module class, a lot of methods are currently reaching into each others internals in ways they really shouldn't. I have now fixed so that doing the actual simulation in the simulated_learning_module is not done by way of the study.