Implemented correct_onwards
The participants class now has the attribute correct_onwards, which states for each session how many questions were correctly answered in that session and every subsequent session. This is meant to be used by the learning_module class. Instead of looking at the number of correct answers in the final session it should look at how quickly learning happens over the course of the learning module.