Partly incorporated participants in learning_module class
I am working on making the learning_module class make use of the participants class, since that will allow us to make an analysis that makes use of all of the data. Right now I'm at a convenient midway point, where the information in the participants is available to the learning_module, but it analyses it in the old way, meaning that I can compare my results to the old ones to verify that I haven't broken anything.