Implemented ranking
Participants are now ranked by how well they did over the entire course of the learning module. This means that participants who had the same score on the final session now have an internal ranking based on how well they did on previous sessions, but it is not possible to beat somebody who did better on a particular session by having done better on a prior session.