Implemented use of ranking in median test
When comparing different backgrounds and manipulations using the median test, we now look at the results for the entire module. That is, two participants who both got X answers correct in the final session might be ranked differently if they got different numbers right on the previous session, or the one before that. I have not implemented the ranking in the boundary tests, since I do not know that there is any boundary that makes sense there.