Made workaround for bug or undocumented feature in numpy
For some reason, applying powers to ndarrays can cause them to suddenly change type if the exponent is large enough. I added a workaround to handle that. This said, the code that contains the workaround is not very cleverly thought out to begin with, so I will probably be rewriting it at some point.