Added learned_background
The module now allows for the possibility that some background variables are knowable to the experimenters, but instead of being known ahead of time they are learned in the course of the learning module. This means that when defining manipulation groups, these background variables cannot be taken into account, but it is still possible to do statistical analysis on them. (The statistical analysis is yet to be implemented).