Changed documentation text
I had previously thought of this as simply a proof-of-concept for code that we will later work into the factorial_experiment module. For reasons of division of labour, it looks like it will now be a fairly sizeable module of its own, which will mostly have the purpose of producing plots that demonstrate how our work is going. Because of that, I may rename it into something more suitable.