Made correct_first_try into a DataFrame
Instead of having a dict or a list, I will just use a DataFrame to keep track of which sessions a participant got correctly the first time around. However, note that it currently outputs an annoying memory message even though it works, so I seem to have done things in a silly way.