Implemented reading of multiple sessions
Now that I know better what the data files look like, I was able to implement reading of results from different sessions. (Assuming, of course, that this does not get changed during the course of the project). I also implemented inference of participant IDs from the results file, so it is no longer necessary to input them when instantiating a learning_module object. That said, it probably will be necessary to do it that way in the end, to catch participants who have signed up for the learning module and logged in once, but who have not actually done a single question.