Implemented input/output of data
I am currently working on making this module consistent with the other modules. It will now output data in a format that the sharepoint_robot module can read, and also in a format that Canvas can read. I have also added warnings regarding the data that is in principle sensitive information, such as the passwords and the real name-user ID key.