Removed incorrect text
After some thought I realised that the premise for this whole thing is incorrect. A maximum-likelihood estimate of a parameter is in fact robust to the choice of parametrisation. This is even kind of obvious: The likelihood is the probability of making a certain observation, given that the system you are observing is in a certain state, and that probability does not depend on your parametrisation. Hence, maximum-likelihood methods (and equivalently, maximum-posterior methods with a flat prior) do not need to worry about the parametrisation. The reason I thought otherwise is due to a hasty reading of von Toussaint (2011). That said, I think the experiment is still interesting, so I will continue working on it tomorrow.