Added script for running AIG.
We now have a module which runs AIG on a file of participant data, which is currently assumed to be the digital coaches in the project. I have also changed the demonstration script into a test script, since that is what I have used it for - testing that the module actually seems to work. I have also added an empty directory names Participant_data, for the convenience of whoever runs the script.