Fixed incorrect array transposition
Earlier when I transposed the array 'Results', I missed the fact that there are two arrays of that name, and only one should be transposed. This fixed that mistake. I also added an array slicing to handle the fact that sometimes - at least right now while we are still working on the modules - some skills will have more sessions than the rest of the module.