Changed method name to clarify that raw_analytics is used
Since I am about to create a second way of importing data, I change the name of the original importing function to clarify what type of file is to be used. I will now attempt to add importing of the file Datashop_foo.xml that you can also get out of OLI Torus.