Updated code to reflect algorithm combining Datashop and raw_analytics
After some work it turned out that neither the Datashop file nor the raw_analytics file contain the information that we need. The raw_analytics file contain all the information about student interactions with OLI Torus, but the user IDs are replaced with automatically generated pseudonyms. The Datashop file on the other hand does contain the correct user IDs, but multiple interactions are combined under a single timestamp, with the internal chronological order scrambled. To get around that, we now read both files and use the Datashop file to figure out the mapping between user IDs and pseudonyms, and then take all the data from the raw_analytics file. The algorithm for figuring out the mapping is slightly complex, but boils down to assuming that if each time a user ID has interacting with OLI Torus coincides with some time that a pseudonym has interacted with it, then the ID and the pseudonym are probably the same person. (Note that it does not work the other way around - because the timestamps in the Datashop file have been combined, there are many timestamps in the raw_analytics file that do not have any corresponding timestamp in the Datashop file.)