Algorithm for figuring out mapping pseudonym->ID now uses best match
We now assume that each pseudonym corresponds to the ID that gives the largest number of matching timestamps. We also save three ways of viewing the mapping, named 'mapping', 'mapping_pseudonym_ID' and 'mapping_ID_pseudonym'. The first is a pandas Dataframe and the second two are dicts. However, note that right now it picks out some correspondences where the match_percentage is zero. I will need to fix that.