Optimised error printout
Now error reporting when importing results is optional. Instead of printing every unmatched ID and pseudonym out, only the total numbers are reported. At the same time, lists containing the unmatched IDs and pseudonyms are saved in the learning_module object, so that they can be inspected later if needed.