Now performs cutting my date immediately when loading data
Previously, when importing from Datashop and raw_analytics, the entire contents of the files would be included. Only later on would cuts be made to pick out the entries that fall within the time span of the course. I had a vague idea that it would be good to include all available information to do the matching between pseudonyms and IDs. I then realised that this was a bad idea, since this in any case won't be available once we are running the modules for real. Hence, we now do the cutting immediately and only save the full original data just in case.