Reintroduced filtering by section_slug
There are a couple of anomalies in the data that I believe are due to there being multiple versions of the course module around. I have enabled filtering by the 'Section Slug' column in the raw_analytics file to remove those entries that I believe are from obsolete or experimental versions of the course.