Implemented function for file upload
We now have a function that uploads files to Canvas. The next step will be to ensure that they actually end up where we want them. After all, the goal is to send files to each individual user, giving feedback on their performance. These files would in turn have to be created by the preprocessing module.