Now only created a preliminary SCB file
The file to be delivered to SCB will have to contain a column with the 'personnummer' - which roughly translates as personal identity number - of each participant. I've decided that the preprocessing module should not have access to this information. It belongs in the part of our code that is completely blind to the real-world identities of the participants. Instead, it will make a preliminary file with a column of user names, which will then have to be replaced by personal identity numbers by a module that has that key.