Fixed a few issues in file uploading
When testing out the uploading in larger scale, the following bugs were found and fixed: 1. The hash function occasionally creates user IDs containing /, which is not allowed in filenames on many systems. To get around this, the files named after users now have those replaced with _ instead. 2. When uploading several conversation attachments to Canvas, each one is overwritten by the previous. To get around this, the filename now explicitly mentions the target user instead of simply being 'Återkoppling'. 3. It turns out that the file area is fairly limited. To cut down the file size into a 6th, the EFS logo is no longer included.