Now feedback is never written for a participant who has not finished the module
During testing we ran into the problem that participants who had not yet done the module - or, more confusingly to them, had done it but before the time span that we tested - got feedback telling them they had answered zero questions correctly and advising them to take all of the learning modules to remedy this. Now we make sure that no such feedback is written.