Module now stores who has already received feedback
In the folder of feedback we create a folder named tracking, which contains a list of who has already received feedback. This will contain a file named 'current', listing all who have received feedback so far, as well as files named after the current date (numbered if many are created on the same day) which keep track of who had received feedback by what date.