Now we testing second-best ID-pseudonym match
The algorithm for comparing IDs from the Datashop file and pseudonyms from the raw_analytics file now also checks how best the second-best match was. If the best and second-best are too similar, it will refuse to pick either. It also saves the second-best match percentage in the mapping file, to make visual inspection easier.