Implemented reading only from Datashop file
We can now infer participant results only from the datashop file. Doing this is not good, since that file does not contain time-ordering, meaning that if a person answered correctly on the first try and then checked three alternatives out of curiosity, they show up to us as having made three incorrect tries before getting it right. Unfortunately, this might be the only viable option, since we do not seem to be able to figure out the mapping between the Datashop and raw_analytics files.