Experiment with improving the algorithm.
It seems that timestamps in the Datashop file typically precede those in the raw_analytics file, since the Datashop timestamp simply shows when somebody started looking at a question, while the raw_analytics timestamp shows when they clicked on individual answers. With somebody who reads slowly - or who went up to get a cup of coffee - the timestamp in the raw_analytics file might be five minutes after that in the Datashop file. Currently the code allows several minutes of delay, but even that does not seem to be enough. Writing this message, another algorithm occurred to me: What if the timespan is extended to whenever the participant next started on a question? That is the only endpoint we actually know, but it might be enough.