Changed ordering of answers dict so that time precedes problem name
Previously I ordered this dict so that things would be grouped by problem name. Now they are grouped by the timestamp given in the Datashop file. This gives a clearer reflection of how data is actually ordered in the Datashop file, which also makes it clear why the algorithm that we tried to use unfortunately won't work, but gives some indication that the algorithm I suggested a few commits earlier might work.