Logs date of next activity in Datashop
We want to know within which interval of time an answer must have been given, based on the data in the Datashop file. To do that we check when the user starts on their next activity. Note that this includes things like survey answers, which do not show up in the xml_data_full or xml_data Dataframes.